fix: Add mutex lock for RabbitMQ connection handling

+++ common_lib/src/ut.rs
@@ -1,4 +1,4 @@
-pub fn calc_bucket_name(prefix: &str, protocol: &str, id: u32) -> String {
+pub fn calc_bucket_name(prefix: &str, protocol: &str, id: u64) -> String {
     format!("{}_{}_{}", prefix, protocol, id % 100)
 }
